refactor: extract submission to index
