+++ web/src/routes/index/ChatInput/index.module.css
@@ -1,4 +1,4 @@
-.input-container {
+.inputContainer {
     margin-left: auto;
     margin-right: auto;
     width: 40vw;
@@ -7,7 +7,7 @@
     align-items: center;
 }
 
-.input-text {
+.inputText {
     width: 100%;
     max-height: 10vw;
     padding: 12px;
@@ -21,20 +21,12 @@
     scrollbar-width: thin;
 }
 
-.input-text:focus {
+.inputText:focus {
     outline: none !important;
     border: 1px solid var(--border-color);
 }
 
-.input-helper {
-    font-size: 0.75em;
-    text-align: start;
-    padding-left: 12px;
-    border-bottom-left-radius: 5px;
-    color: var(--text-secondary);
-}
-
-.input-submit-button {
+.inputSubmitButton {
     width: 70px;
     height: 60px;
     margin-left: 2vw;
@@ -45,6 +37,6 @@
     color: var(--text-primary);
 }
 
-.input-submit-button:disabled {
+.inputSubmitButton:disabled {
     background-color: var(--text-secondary);
 }
