+++ web/src/routes/index/index.jsx
@@ -1,15 +1,46 @@
 import "./index.css";
 
 import { useContext } from "react";
+import { useNavigate } from "react-router-dom";
 
 import { UserContext } from "contexts/user";
+import { ConversationContext } from "contexts/conversation";
 import ChatboxHeader from "components/ChatboxHeader";
+import { DEFAULT_CONV_TITLE } from "commons";
 
 import ChatInput from "./ChatInput";
 
 
 const Conversation = () => {
     const { username } = useContext(UserContext);
+    const { dispatch } = useContext(ConversationContext);
+    const navigate = useNavigate();
+
+    const handleSubmit = async (input) => {
+        try {
+          const response = await fetch("/api/conversations", {
+            method: "POST",
+            headers: {
+              "Content-Type": "application/json",
+            },
+            body: JSON.stringify({ title: DEFAULT_CONV_TITLE }),
+          });
+          const conversation = await response.json();
+          const message = {
+            id: crypto.randomUUID(),
+            conversation: conversation.id,
+            from: username,
+            content: input,
+            type: "text",
+          };
+          sessionStorage.setItem(`init-msg:${conversation.id}`, JSON.stringify(message));
+          dispatch({ type: "added", conv: conversation });
+          navigate(`/conversations/${conversation.id}`);
+        } catch (error) {
+          console.error('Error creating conversation:', error);
+        }
+      };
+
     return (
         <>
             <ChatboxHeader />
@@ -19,7 +50,7 @@ const Conversation = () => {
             </div>
             {/* TODO: add some examples here? */}
             <div className="input-bottom">
-                <ChatInput />
+                <ChatInput onSubmit={handleSubmit} />
                 <div className="footer">Chatbot can make mistakes. Consider checking important information.</div>
             </div>
         </>
